+++ CHAPTER6/キャラクタ/__init__.py
@@ -4,8 +4,8 @@
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #from .自機      import class自機
 from . import 自機
-from . import 標的
+from . import 敵機
 from . import アイテム
 
-from .弾        import class弾
-from .爆発.爆発 import class爆発
+from .弾    import *
+from .爆発  import *
