+++ CHAPTER6/キャラクタ/アイテム/__init__.py
@@ -2,5 +2,5 @@
 #┃技術評論社 ゲームで学ぶPython！ CHAPTER6:MAGA WING
 #┃キャラクター・モジュール
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from .アイテム  import classアイテム
-from .アイテム  import ID as アイテムID
+from .main  import 本体 as アイテム
+from .main  import ID   as アイテムID
