+++ CHAPTER6/キャラクタ/アイテム/main.py
@@ -1,0 +1,91 @@
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃キャラクタ：アイテム
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+from  .ID       import ID
+from  .FN移動   import 移動クラス
+from  .FN衝突   import 衝突クラス
+from  .FN描画   import 描画クラス
+
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃仕様
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+class 仕様クラス:
+
+    #□仕様データベース
+    #・難易度制限   ： 難易度がこの数値異常で出現する 
+    #・持続時間     ： 効果が続く時間(単位：秒)
+    DB = {
+        ID.速度UP	        : (  0,  0,),
+        ID.弾薬回復	        : (  0,  0 ),
+        ID.シールド回復     : (  0,  0 ),
+        ID.発射_連射   	    : (  0,  0 ),
+        ID.発射_弾数UP      : (  0,  0 ),
+        ID.ダメージ無し	    : (  0, 10 ),
+        ID.ダメージ半分	    : (  0, 10 ),
+        ID.ダメージ吸収     : (  0, 10 ),
+        ID.発射_貫通弾	    : (  0, 10 ),
+        ID.発射_上下方向    : (  0, 10 ),
+        ID.発射_左右方向    : (  0, 10 ),
+        ID.発射_4方向   	: (  0, 10 ),
+        ID.発射_8方向   	: (  0, 10 )}
+
+    def __init__(self, 引数_種類ID): #① 生成するアイテムの種類ID
+        #┬
+        #□仕様データ
+        self.種類ID     = 引数_種類ID
+        self.難易度制限 = self.DB[ 引数_種類ID ][0]
+        self.持続時間   = self.DB[ 引数_種類ID ][1]
+        #┴
+
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃情報
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+class 情報クラス:
+
+    def __init__(self   ,
+            引数_X座標  ,   #① 存在位置(X座標)
+            引数_Y座標  ):  #② 存在位置(Y座標)
+		#┬
+        #□動的データ
+        self.X          = 引数_X座標
+        self.Y          = 引数_Y座標
+        #┴
+
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃メイン
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+class 本体:
+    #┌───────────────────────────────────
+    #│初期化
+    #└───────────────────────────────────
+    def __init__(self   ,
+            引数_ゲーム ,	#① ゲームのオブジェクト
+            引数_X座標  ,   #② 位置(X座標)
+            引数_Y座標  ,	#③ 位置(Y座標)
+            引数_種類ID ):  #④ 出現する種類のID
+
+		#┬
+        #〇リファレンスを用意する
+        self.格納先 = 引数_ゲーム.objアイテム
+        #│
+        #〇└┐データセットを用意する
+            #≫仕様
+            #≫情報
+        self.仕様   = 仕様クラス(引数_種類ID)
+        self.情報   = 情報クラス(引数_X座標, 引数_Y座標)
+        #│ｓｓ
+        #〇処理セットを用意する
+        self.FN移動 = None
+        self.FN衝突 = None
+        self.FN特殊 = None
+        self.FN発射 = None
+        self.FN描画 = None
+        #│
+        #≫処理セットをオブジェクト化する
+        移動クラス(self)
+        衝突クラス(self)
+        描画クラス(self)
+        #│
+        #●格納先インスタンスにオブジェクトを追加する
+        self.格納先.append(self)
+        #┴
