+++ CHAPTER6/キャラクタ/弾/__init__.py
@@ -1,5 +1,4 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃技術評論社 ゲームで学ぶPython！ CHAPTER6:MAGA WING
-#┃キャラクター・モジュール
+#┃キャラクター：弾
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from .弾  import class弾
+from .main  import 本体 as 弾
