+++ CHAPTER6/キャラクタ/弾/main.py
@@ -1,40 +1,28 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃キャラクタ：アイテム
+#┃キャラクタ：アイテム　※自機・敵機で共用する
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from  .ID       import ID
-from  .FN移動   import 移動クラス
-from  .FN衝突   import 衝突クラス
-from  .FN描画   import 描画クラス
+from    処理    import class所有者ID as 所有者ID
+from    .FN移動 import 移動クラス
+from    .FN衝突 import 衝突クラス
+from    .FN描画 import 描画クラス
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 class 仕様クラス:
 
-    #□仕様データベース
-    #・難易度制限   ： 難易度がこの数値異常で出現する 
-    #・持続時間     ： 効果が続く時間(単位：秒)
-    DB = {
-        ID.速度UP	        : (  0,  0,),
-        ID.弾薬回復	        : (  0,  0 ),
-        ID.シールド回復     : (  0,  0 ),
-        ID.発射_連射   	    : (  0,  0 ),
-        ID.発射_弾数UP      : (  0,  0 ),
-        ID.ダメージ無し	    : (  0, 10 ),
-        ID.ダメージ半分	    : (  0, 10 ),
-        ID.ダメージ吸収     : (  0, 10 ),
-        ID.発射_貫通弾	    : (  0, 10 ),
-        ID.発射_上下方向    : (  0, 10 ),
-        ID.発射_左右方向    : (  0, 10 ),
-        ID.発射_4方向   	: (  0, 10 ),
-        ID.発射_8方向   	: (  0, 10 )}
+    def __init__(self,引数_所有者ID):   #① 弾の所有者
 
-    def __init__(self, 引数_種類ID): #① 生成するアイテムの種類ID
         #┬
-        #□仕様データ
-        self.種類ID     = 引数_種類ID
-        self.難易度制限 = self.DB[ 引数_種類ID ][0]
-        self.持続時間   = self.DB[ 引数_種類ID ][1]
+        #□仕様データ(共通)
+        self.種類ID     = None
+        self.所有者ID   = 引数_所有者ID
+        #│
+        #□所有者別データ
+        if self.所有者ID == 所有者ID.自機:
+            self.撃墜可否 = False
+        else:
+            self.撃墜可否 = True
         #┴
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
@@ -42,36 +30,40 @@ class 仕様クラス:
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 class 情報クラス:
 
-    def __init__(self   ,
-            引数_X座標  ,   #① 存在位置(X座標)
-            引数_Y座標  ):  #② 存在位置(Y座標)
-		#┬
-        #□動的データ
+    def __init__(self,
+            引数_X座標      ,   #① 存在位置(X座標)
+            引数_Y座標      ):  #② 存在位置(Y座標)
+
         self.X          = 引数_X座標
         self.Y          = 引数_Y座標
-        #┴
+        self.シールド   = 0
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class classアイテム:
+class 本体:
     #┌───────────────────────────────────
     #│初期化
     #└───────────────────────────────────
-    def __init__(self   ,
-            引数_ゲーム ,	#① ゲームのオブジェクト
-            引数_X座標  ,   #② 位置(X座標)
-            引数_Y座標  ,	#③ 位置(Y座標)
-            引数_種類ID ):  #④ 出現する種類のID
+    def __init__(self       ,
+            引数_ゲーム     ,   #① ゲームのオブジェクト
+            引数_所有者ID   ,   #② 弾の所有者
+            引数_X座標      ,	#③ 存在位置(X座標)
+            引数_Y座標      ,	#④ 存在位置(Y座標)
+            引数_発射角度   ,   #⑤ 発射する方向(角度)
+            引数_移動速度   ):  #⑥ 移動速度(XYベクトル)
 
 		#┬
         #〇リファレンスを用意する
-        self.格納先 = 引数_ゲーム.objアイテム
+        if 引数_所有者ID == 所有者ID.自機:
+            self.格納先 = 引数_ゲーム.obj弾_自機
+        else:
+            self.格納先 = 引数_ゲーム.obj弾_敵機
         #│
         #〇└┐データセットを用意する
             #≫仕様
             #≫情報
-        self.仕様   = 仕様クラス(引数_種類ID)
+        self.仕様   = 仕様クラス(引数_所有者ID)
         self.情報   = 情報クラス(引数_X座標, 引数_Y座標)
         #│
         #〇処理セットを用意する
@@ -82,8 +74,8 @@ class classアイテム:
         self.FN描画 = None
         #│
         #≫処理セットをオブジェクト化する
-        移動クラス(self)
-        衝突クラス(self)
+        移動クラス(self, 引数_発射角度, 引数_移動速度)
+        衝突クラス(self, 引数_ゲーム)
         描画クラス(self)
         #│
         #●格納先インスタンスにオブジェクトを追加する
