+++ CHAPTER6/キャラクタ/敵機/FN描画.py
