+++ CHAPTER6/キャラクタ/敵機/FN移動.py
