+++ CHAPTER6/キャラクタ/敵機/FN衝突.py
@@ -1,8 +1,8 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃キャラクタ：標的：衝突機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from    処理        import class所有者ID    as 所有者ID 
-from    ..爆発.爆発 import class爆発
+from 処理    import class所有者ID    as 所有者ID 
+from ..爆発  import *
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
@@ -79,7 +79,7 @@ class 衝突クラス:
         #│├┐（爆破指定がある場合）
             #↓
             #●爆発を生成する ※音あり
-            class爆発(格納先, 所有者, x + 4, y + 4, 0, 0, r, True)
+            爆発(格納先, 所有者, x + 4, y + 4, 0, 0, r, True)
             #┴
         #　└┐（その他）
             #┴
