+++ CHAPTER6/キャラクタ/敵機/ID.py
