+++ CHAPTER6/キャラクタ/敵機/main.py
@@ -46,7 +46,7 @@ class 情報クラス:
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class class標的:
+class 本体:
     #┌───────────────────────────────────
     #│初期化
     #└───────────────────────────────────
