+++ CHAPTER6/キャラクタ/爆発/__init__.py
@@ -1,4 +1,4 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃キャラクター・モジュール
+#┃キャラクター：爆発
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from .爆発  import class爆発
+from .main  import 本体 as 爆発
