+++ CHAPTER6/キャラクタ/爆発/main.py
@@ -1,9 +1,8 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃キャラクタ：アイテム　※自機・敵機で共用する
+#┃キャラクタ：爆発　※自機・敵機で共用する
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from    処理    import class所有者ID as 所有者ID
+import  pyxel
 from    .FN移動 import 移動クラス
-from    .FN衝突 import 衝突クラス
 from    .FN描画 import 描画クラス
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
@@ -11,19 +10,9 @@ from    .FN描画 import 描画クラス
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 class 仕様クラス:
 
-    def __init__(self,引数_所有者ID):   #① 弾の所有者
+    def __init__(self, 引数_所有者ID): #① 爆発の所有者
 
-        #┬
-        #□仕様データ(共通)
-        self.種類ID     = None
-        self.所有者ID   = 引数_所有者ID
-        #│
-        #□所有者別データ
-        if self.所有者ID == 所有者ID.自機:
-            self.撃墜可否 = False
-        else:
-            self.撃墜可否 = True
-        #┴
+        self.所有者ID = 引数_所有者ID
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃情報
@@ -31,34 +20,33 @@ class 仕様クラス:
 class 情報クラス:
 
     def __init__(self,
-            引数_X座標      ,   #① 存在位置(X座標)
-            引数_Y座標      ):  #② 存在位置(Y座標)
+            引数_X座標  ,   #① 存在位置(X座標)
+            引数_Y座標  ):  #② 存在位置(Y座標)
 
-        self.X          = 引数_X座標
-        self.Y          = 引数_Y座標
-        self.シールド   = 0
+        self.X      = 引数_X座標
+        self.Y      = 引数_Y座標
+        self.半径   = 0
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class class弾:
+class 本体:
     #┌───────────────────────────────────
     #│初期化
     #└───────────────────────────────────
-    def __init__(self       ,
-            引数_ゲーム     ,   #① ゲームのオブジェクト
-            引数_所有者ID   ,   #② 弾の所有者
-            引数_X座標      ,	#③ 存在位置(X座標)
-            引数_Y座標      ,	#④ 存在位置(Y座標)
-            引数_発射角度   ,   #⑤ 発射する方向(角度)
-            引数_移動速度   ):  #⑥ 移動速度(XYベクトル)
-
+    def __init__(self,
+            引数_格納先     , #① 生成先のオブジェクト
+            引数_所有者ID   , #② 爆発の所有者
+            引数_X座標      , #③ 存在位置(X座標)
+            引数_Y座標      , #④ 存在位置(Y座標)
+            引数_速度X      , #⑤ 発射する方向(角度)
+            引数_速度Y      , #⑥ 移動速度(XYベクトル)
+            引数_最大半径   , #⑦ 爆発の範囲(半径)
+            引数_爆発音有無 ):#⑧ 爆発音の有無
 		#┬
-        #〇リファレンスを用意する
-        if 引数_所有者ID == 所有者ID.自機:
-            self.格納先 = 引数_ゲーム.obj弾_自機
-        else:
-            self.格納先 = 引数_ゲーム.obj弾_敵機
+        #□本体オブジェクト(参照用)
+        self.格納先 = 引数_格納先
+            #┴
         #│
         #〇└┐データセットを用意する
             #≫仕様
@@ -74,10 +62,12 @@ class class弾:
         self.FN描画 = None
         #│
         #≫処理セットをオブジェクト化する
-        移動クラス(self, 引数_発射角度, 引数_移動速度)
-        衝突クラス(self, 引数_ゲーム)
+        移動クラス(self,引数_速度X, 引数_速度Y, 引数_最大半径)
         描画クラス(self)
         #│
-        #●格納先インスタンスにオブジェクトを追加する
+        #●本体で用意したインスタンスに生成
         self.格納先.append(self)
+        #│
+        #○爆発音を鳴らす
+        if 引数_爆発音有無: pyxel.play( 0, 2, resume = True ) 
         #┴
