--- CHAPTER6/キャラクタ/爆発/爆発.py
@@ -1,73 +1,0 @@
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃キャラクタ：爆発　※自機・敵機で共用する
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-import  pyxel
-from    .FN移動 import 移動クラス
-from    .FN描画 import 描画クラス
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃仕様
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 仕様クラス:
-
-    def __init__(self, 引数_所有者ID): #① 爆発の所有者
-
-        self.所有者ID = 引数_所有者ID
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃情報
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 情報クラス:
-
-    def __init__(self,
-            引数_X座標  ,   #① 存在位置(X座標)
-            引数_Y座標  ):  #② 存在位置(Y座標)
-
-        self.X      = 引数_X座標
-        self.Y      = 引数_Y座標
-        self.半径   = 0
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃メイン
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class class爆発:
-    #┌───────────────────────────────────
-    #│初期化
-    #└───────────────────────────────────
-    def __init__(self,
-            引数_格納先     , #① 生成先のオブジェクト
-            引数_所有者ID   , #② 爆発の所有者
-            引数_X座標      , #③ 存在位置(X座標)
-            引数_Y座標      , #④ 存在位置(Y座標)
-            引数_速度X      , #⑤ 発射する方向(角度)
-            引数_速度Y      , #⑥ 移動速度(XYベクトル)
-            引数_最大半径   , #⑦ 爆発の範囲(半径)
-            引数_爆発音有無 ):#⑧ 爆発音の有無
-		#┬
-        #□本体オブジェクト(参照用)
-        self.格納先 = 引数_格納先
-            #┴
-        #│
-        #〇└┐データセットを用意する
-            #≫仕様
-            #≫情報
-        self.仕様   = 仕様クラス(引数_所有者ID)
-        self.情報   = 情報クラス(引数_X座標, 引数_Y座標)
-        #│
-        #〇処理セットを用意する
-        self.FN移動 = None
-        self.FN衝突 = None
-        self.FN特殊 = None
-        self.FN発射 = None
-        self.FN描画 = None
-        #│
-        #≫処理セットをオブジェクト化する
-        移動クラス(self,引数_速度X, 引数_速度Y, 引数_最大半径)
-        描画クラス(self)
-        #│
-        #●本体で用意したインスタンスに生成
-        self.格納先.append(self)
-        #│
-        #○爆発音を鳴らす
-        if 引数_爆発音有無: pyxel.play( 0, 2, resume = True ) 
-        #┴
