+++ CHAPTER6/キャラクタ/自機/FN発射.py
@@ -4,8 +4,8 @@
 import  pyxel
 from    共通	    import class入力操作    as 入力 
 from    処理        import class所有者ID    as 所有者ID 
-from    ..弾        import class弾
-from    .FN特殊 import 効果ID
+from    ..弾        import *
+from    .FN特殊     import 効果ID
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
@@ -175,7 +175,7 @@ class 発射クラス:
         #○弾を生成する
         x = self.基底.情報.X + argオフセットX
         y = self.基底.情報.Y + argオフセットY
-        class弾(
+        弾(
                 self.基底.GAME  ,
                 所有者ID.自機   ,
                 x, y, -90, 5    )
