+++ CHAPTER6/キャラクタ/自機/FN衝突.py
@@ -1,9 +1,8 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃キャラクタ：自機：衝突機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-import  pyxel
-from    処理    import class所有者ID    as 所有者ID 
-from    ..爆発.爆発  import class爆発
+from 処理        import class所有者ID    as 所有者ID 
+from ..爆発.main import 本体             as 爆発
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
@@ -88,7 +87,7 @@ class 衝突クラス:
         #　├┐（爆破指定がある場合）
             #↓
             #●爆発を生成する ※音あり
-            class爆発(格納先, 所有者, x + 4, y + 4, 0, 0, r, True)
+            爆発(格納先, 所有者, x + 4, y + 4, 0, 0, r, True)
             #┴
         #　└┐（その他）
             #┴
@@ -103,6 +102,6 @@ class 衝突クラス:
             self.基底.GAME.obj自機 = None
             #│
             #●追加の爆発を生成する
-            class爆発(格納先, 所有者, x + 10, y + 10, 0, 0, r, True)
-            class爆発(格納先, 所有者, x -  2, y -  2, 0, 0, r, True)
+            爆発(格納先, 所有者, x + 10, y + 10, 0, 0, r, True)
+            爆発(格納先, 所有者, x -  2, y -  2, 0, 0, r, True)
         #┴　┴
