+++ CHAPTER6/処理/main.py
@@ -8,9 +8,11 @@
 #┃Ⅰ.インポート
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import pyxel
+
 from .ID            import *
+
 from シーン         import *
-from 演出	        import class背景
+from 背景	        import *
 
 # 更新処理リストの要素
 from .更新1_移動    import 移動クラス
@@ -41,7 +43,7 @@ class 情報クラス:
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class classGame:
+class 本体:
 	#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 	#┃０．初期化処理 
     #┠───────────────────────────────────
@@ -87,7 +89,7 @@ class classGame:
             シーンID.終了       : class終了     (self)}
             #│
             #≫背景をオブジェクト化する
-        self.obj背景        = class背景(self)
+        self.obj背景 = 背景(self)
             #┴
         #│
         #≫シーンを切替える(タイトル)
