+++ CHAPTER6/処理/更新3_衝突.py
@@ -1,10 +1,8 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃更新機能：衝突
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-import  pyxel
-from    共通		        import class汎用    as 共通
-from    キャラクタ.アイテム import アイテムID
-from    キャラクタ.標的     import 敵機ID
+from    共通		    import class汎用    as 共通
+from    キャラクタ.敵機 import 敵機ID
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
