+++ CHAPTER6/処理/更新4_出現.py
@@ -3,8 +3,8 @@
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import  pyxel
 from    シーン              import シーンID
-from    キャラクタ.標的     import class標的, 敵機ID
-from    キャラクタ.アイテム import classアイテム, アイテムID
+from    キャラクタ.敵機     import *
+from    キャラクタ.アイテム import *
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
@@ -13,9 +13,9 @@ class 仕様:
     #┬
     #□出現間隔
     機雷    = 200
-    弾薬箱  = 400
-    救急箱  = 700
-    連射    = 300
+    弾薬箱  = 200
+    救急箱  = 300
+    連射    = 100
     #┴
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
@@ -67,7 +67,7 @@ class 出現クラス:
         X座標   = pyxel.rndi(0, pyxel.width - 8)
         種類ID  = pyxel.rndi(敵機ID.戦闘機1, 敵機ID.戦闘機3)
         難易度  = self.基底.情報.難易度
-        class標的(self.基底, X座標, -8, 種類ID, 難易度)
+        敵機(self.基底, X座標, -8, 種類ID, 難易度)
         #┴
 	#────────────────────────────────────
     def Fn機雷(self):
@@ -82,7 +82,7 @@ class 出現クラス:
         X座標   = pyxel.rndi(0, pyxel.width - 8)
         種類ID  = 敵機ID.機雷
         難易度  = pyxel.rndi(3, 5) 
-        class標的(self.基底, X座標, -8, 種類ID, 難易度)
+        敵機(self.基底, X座標, -8, 種類ID, 難易度)
         #┴
 	#────────────────────────────────────
     def Fnアイテム(self, 引数_間隔, 引数_種類ID):
@@ -95,5 +95,5 @@ class 出現クラス:
         #│
         #●救急箱を生成する
         X座標 = pyxel.rndi(0, pyxel.width - 8)
-        classアイテム(self.基底, X座標, -8, 引数_種類ID)
+        アイテム(self.基底, X座標, -8, 引数_種類ID)
        #┴
