+++ CHAPTER6/背景/FN描画.py
