+++ CHAPTER6/背景/FN移動.py
