+++ CHAPTER6/背景/__init__.py
@@ -1,5 +1,4 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃技術評論社 ゲームで学ぶPython！ CHAPTER6:MAGA WING
-#┃演出モジュール
+#┃背景
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from .背景 import class背景
+from .main import 本体 as 背景
