+++ CHAPTER6/背景/main.py
@@ -1,68 +1,66 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃キャラクタ：自機
+#┃演出：背景
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import  pyxel
-from    .FN移動     import 移動クラス
-from    .FN衝突     import 衝突クラス
-from    .FN特殊     import 特殊クラス
-from    .FN発射     import 発射クラス
-from    .FN描画     import 描画クラス
+from    .FN移動 import 移動クラス
+from    .FN描画 import 描画クラス
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 class 仕様クラス:
-    pass
+
+    星数 = 30
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃情報
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 class 情報クラス:
 
-    def __init__(self       ,
-            引数_X座標      ,   #① 存在位置(X座標)
-            引数_Y座標      ):  #② 存在位置(Y座標)
+    def __init__(self):
 
-        self.X          = 引数_X座標
-        self.Y          = 引数_Y座標
-        self.シールド   = int(pyxel.width * 0.5)
+        self.星 = []        # 星リスト
+
+        #◎└┐すべての星を追加する
+        for cnt in range(仕様クラス.星数):
+            #│＼（すべての星を処理し終えた場合）
+            #│ ▼繰り返し処理を抜ける
+            #│
+            #○位置をセットする
+            座標X = pyxel.rndi(0, pyxel.width  - 1  )  
+            座標Y = pyxel.rndi(0, pyxel.height - 1  )  
+            #│
+            #○速度をセットする
+            速度Y = pyxel.rndf(0.5, 0.8             ) 
+            #│
+            #○星を追加する
+            self.星.append((座標X, 座標Y, 速度Y))
+        #┴　┴
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class class自機:
+class 本体:
     #┌───────────────────────────────────
     #│初期化
     #└───────────────────────────────────
-    def __init__(self   ,
-            引数_ゲーム ,	#① ゲームのオブジェクト
-            引数_X座標  ,	#② 存在位置(X座標)
-            引数_Y座標  ):  #③ 存在位置(Y座標)
-		#┬
-        #〇リファレンスを用意する
-        self.GAME   = 引数_ゲーム
-        #self.格納先 = 引数_ゲーム.obj自機  #★★★なぜか機能しない★★★
+    def __init__(self, 引数_生成先): #① 生成先のオブジェクト
+
+        #┬
+        #□生成先のオブジェクト
+        self.GAME	= 引数_生成先
         #│
         #〇└┐データセットを用意する
             #≫仕様
             #≫情報
-        self.仕様   = None
-        self.情報   = 情報クラス(引数_X座標, 引数_Y座標)
+        self.仕様   = 仕様クラス()
+        self.情報   = 情報クラス()
         #│
         #〇処理セットを用意する
         self.FN移動 = None
-        self.FN衝突 = None
-        self.FN特殊 = None
-        self.FN発射 = None
         self.FN描画 = None
         #│
         #≫処理セットをオブジェクト化する
         移動クラス(self)
-        衝突クラス(self, 引数_ゲーム)
-        特殊クラス(self)
-        発射クラス(self, 引数_ゲーム)
         描画クラス(self)
-        #│
-        #●本体で用意したインスタンスに生成
-        引数_ゲーム.obj自機  = self
         #┴
